refactor(db): extract checkpoint capture and packing
